Greed & GameHandler
There were null exceptions with the resultText so just put it inside an if statement to check if it's null before trying to access it in Start()

Also, made it so the play button in the menu scene goes to level_select.

Greed, if you win, goe to level_select, and if you lose, to the lose screen.